test(containers): fold the undefined-name checks into the mirrored probe test
PS-204 orphan-test-file, and the auditor was right:

    tests/.../containers/test_probe_has_no_undefined_names.py: no matching src
    file (orphan test); mirror dir `src/scitex_agent_container/containers`
    contains: sif_symbol_probe.py — none match expected
    `probe_has_no_undefined_names.py`

I had named the file after the CHECK rather than after the thing under test.
The repo's convention is that a test file mirrors a source file, and the
checks belong to sif_symbol_probe.py, so they belong in
test_sif_symbol_probe.py. Merged there and the orphan deleted.

The merge is also better than what I wrote. The existing file derives PROBE
from the INSTALLED package (`Path(scitex_agent_container.__file__).parent /
"containers"`), where mine walked `parents[3]` from the test file — the
former follows the package wherever it is installed, the latter breaks the
moment the tree is laid out differently. EMBEDS now hangs off the same
PROBE.parent.

Nothing was dropped: all three checks survive (AST bind-vs-load on the real
probe, the self-check feeding it the exact #1072 shape, and the token check
over the three embedded heredoc copies), and they sit alongside the existing
checks they complement — test_probe_imports_scitex_cards asserts the symbol
gate is PRESENT, these assert it is not referencing anything that does not
exist.

    ruff check   All checks passed
    pytest       11 passed in the containers/ test dir

Remaining audit findings on this branch are pre-existing repo-level ones and
are NOT addressed here: §4b (help is free-form rather than CliHelp), §10w
(import-budget could not measure — bare-interpreter baseline 141ms exceeds
the 100ms floor), §13 (the `skills` command sits at top level instead of
under `dev`), and the already-masked PS-226. They fail on develop too and
each deserves its own change rather than being smuggled into a probe fix.
