+++ tests/scitex_agent_container/containers/test_sif_symbol_probe.py
@@ -13,6 +13,8 @@ from __future__ import annotations
 import ast
 from pathlib import Path
 
+import pytest
+
 import scitex_agent_container
 
 PROBE = (
@@ -126,3 +128,126 @@ def test_probe_fails_loud_on_a_bad_sif() -> None:
     exits_nonzero = _has_nonzero_sys_exit(source)
     # Assert
     assert exits_nonzero
+
+
+# ---------------------------------------------------------------------------
+# UNDEFINED-NAME CHECKS (added 2026-08-16 after sac #1072)
+#
+# #1072 deleted the ``scitex_todo`` shim and removed every ``import
+# scitex_todo`` line. It did NOT remove the line that USES the name::
+#
+#     import scitex_cards
+#     if scitex_todo is not scitex_cards:      # <- NameError at runtime
+#         sys.exit(1)
+#
+# The probe sys.exit(1)s the bake, so it aborted with ``NameError: name
+# 'scitex_todo' is not defined`` — which reads like "the shim is missing from
+# the image" rather than "the probe is broken". scitex-cards found it by
+# EXECUTING the lines; the suite stayed green because nothing executes the
+# probe.
+#
+# WHY THE ORIGINAL VERIFICATION MISSED IT: it counted anchored
+# ``^\s*(import|from)\s+scitex_todo`` lines and correctly got zero. That
+# answers "is the import gone", never "is the NAME gone" — and the surviving
+# reference was not an import, so the anchor could not see it.
+# ---------------------------------------------------------------------------
+
+import builtins  # noqa: E402
+
+#: The bake recipes embed the same probe as a heredoc, so they cannot be
+#: parsed as Python and get a token check instead.
+EMBEDS = (
+    PROBE.parent / "apptainer-base.def",
+    PROBE.parent / "apptainer-scitex.def",
+    PROBE.parent / "spartan-sif-bake.sh",
+)
+
+
+def _module_level_bound_names(tree: ast.Module) -> set[str]:
+    """Names bound at module level: imports, assignments, defs, args, excepts."""
+    bound: set[str] = set()
+    for node in ast.walk(tree):
+        if isinstance(node, ast.Import):
+            for a in node.names:
+                bound.add(a.asname or a.name.split(".")[0])
+        elif isinstance(node, ast.ImportFrom):
+            for a in node.names:
+                bound.add(a.asname or a.name)
+        elif isinstance(node, (ast.FunctionDef, ast.AsyncFunctionDef, ast.ClassDef)):
+            bound.add(node.name)
+        elif isinstance(node, ast.Name) and isinstance(node.ctx, (ast.Store, ast.Del)):
+            bound.add(node.id)
+        elif isinstance(node, ast.arg):
+            bound.add(node.arg)
+        elif isinstance(node, ast.ExceptHandler) and node.name:
+            bound.add(node.name)
+    return bound
+
+
+def _loaded_names(tree: ast.Module) -> set[str]:
+    return {
+        n.id
+        for n in ast.walk(tree)
+        if isinstance(n, ast.Name) and isinstance(n.ctx, ast.Load)
+    }
+
+
+def _undefined(source: str) -> set[str]:
+    tree = ast.parse(source)
+    return _loaded_names(tree) - _module_level_bound_names(tree) - set(dir(builtins))
+
+
+def test_probe_loads_no_name_it_never_binds() -> None:
+    """The exact #1072 defect: a name used but never imported or assigned."""
+    # Arrange
+    source = _probe_source()
+    # Act
+    undefined = _undefined(source)
+    # Assert
+    assert not undefined, (
+        f"{PROBE.name} references name(s) it never binds: {sorted(undefined)} — "
+        "a runtime NameError inside a gate that sys.exit(1)s, so the bake "
+        "aborts with a message about the wrong thing."
+    )
+
+
+def test_the_undefined_name_check_still_catches_the_1072_shape() -> None:
+    """A gate I cannot show failing is not a gate.
+
+    Feeds the checker the exact code #1072 shipped and requires it to object,
+    so a checker that silently stopped working cannot look like a clean repo.
+    """
+    # Arrange
+    broken = (
+        "import sys\n"
+        "import scitex_cards\n"
+        "if scitex_todo is not scitex_cards:\n"
+        "    sys.exit(1)\n"
+    )
+    # Act
+    undefined = _undefined(broken)
+    # Assert
+    assert undefined == {"scitex_todo"}, (
+        f"the undefined-name check no longer detects the #1072 defect; got {undefined}"
+    )
+
+
+@pytest.mark.parametrize("path", EMBEDS, ids=lambda p: p.name)
+def test_bake_recipes_do_not_use_the_deleted_name(path) -> None:
+    """The embedded probe copies must not mention the deleted module either.
+
+    Comments may still explain the history — that is documentation, not a
+    reference the interpreter will try to resolve.
+    """
+    # Arrange
+    lines = path.read_text(encoding="utf-8").splitlines()
+    # Act
+    offenders = [
+        f"{path.name}:{n}: {ln.strip()}"
+        for n, ln in enumerate(lines, 1)
+        if "scitex_todo" in ln.split("#", 1)[0]
+    ]
+    # Assert
+    assert not offenders, (
+        "deleted module referenced in executable position:\n" + "\n".join(offenders)
+    )
